{{cookiecutter.repo_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.repo_name}}"
version = "0.1.0"
description = "Restful api scaffold created by rest-django-cookiecutter."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.2"
django-environ = "^0.4.5"
bcrypt = "^3.1"
djangorestframework = "^3.10"
celery = "^4.3"
django-cors-headers = "^3.1"

[tool.poetry.dev-dependencies]
pre-commit = "^1.18"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
